+++ src/components/ui/InputText/InputText.tsx
@@ -1,5 +1,6 @@
 import { Input as UIInput, Label as UILabel } from '@faststore/ui'
-import { useEffect, useRef, useState } from 'react'
+import type { MutableRefObject } from 'react'
+import { useEffect, useState } from 'react'
 import Button from 'src/components/ui/Button'
 import ButtonIcon from 'src/components/ui/Button/ButtonIcon'
 import Icon from 'src/components/ui/Icon'
@@ -18,6 +19,10 @@ type DefaultProps = {
    * The error message is displayed when an error occurs.
    */
   errorMessage?: string
+  /**
+   * Component's ref.
+   */
+  inputRef?: MutableRefObject<HTMLInputElement | null>
 }
 
 type ActionableInputText =
@@ -53,12 +58,11 @@ const InputText = ({
   onSubmit,
   placeholder = ' ', // initializes with an empty space to style float label using `placeholder-shown`
   value,
+  inputRef,
   ...otherProps
 }: InputTextProps) => {
   const [inputValue, setInputValue] = useState<string>((value as string) ?? '')
   const [hasError, setHasError] = useState<boolean>(!!errorMessage)
-
-  const inputRef = useRef<HTMLInputElement>(null)
 
   useEffect(() => {
     errorMessage && setHasError(true)
@@ -66,7 +70,7 @@ const InputText = ({
 
   const onClear = () => {
     setInputValue('')
-    inputRef.current?.focus()
+    inputRef?.current?.focus()
   }
 
   return (